Refactored the logging into its own class.  Added development Firebase project for ease of testing.
