+++ navigation/ios_apps/ClewApp/Clew/GoogleService-Info_dev.plist
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>748556153769-ihpe4bsdjrh6mjqtepu3rcc59hv4a55u.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.748556153769-ihpe4bsdjrh6mjqtepu3rcc59hv4a55u</string>
+	<key>API_KEY</key>
+	<string>AIzaSyCNXpIye8PEpxaI09C8A1ROiUmnU35dLZo</string>
+	<key>GCM_SENDER_ID</key>
+	<string>748556153769</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>edu.occam.Clew-Dev</string>
+	<key>PROJECT_ID</key>
+	<string>clew-dev</string>
+	<key>STORAGE_BUCKET</key>
+	<string>clew-dev.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:748556153769:ios:66bcd97155180629</string>
+	<key>DATABASE_URL</key>
+	<string>https://clew-dev.firebaseio.com</string>
+</dict>
+</plist>
