+++ navigation/ios_apps/ClewApp/Clew/Utility.swift
@@ -12,17 +12,8 @@ func roundToTenths(_ n: Float) -> Float {
     return roundf(10 * n)/10
 }
 
-func roundToThousandths(_ n: Double) -> Double {
-    return round(1000 * n)/1000
-}
 
-func round10k(_ n: Float) -> Float {
-    // round Float to then-thousandths
-    return roundf(10000 * n)/10000
-}
-
-
-/// Modulus function (swift's % operator is a remainder function that doens't work properly for negatie numbers
+/// Modulus function (swift's % operator is a remainder function that doesn't work properly for negative numbers
 ///
 /// - Parameters:
 ///   - a: the number
